fix(serve): compile the unix listener only where unix sockets exist
Windows has no unix domain sockets, so `tokio::net::UnixListener` is gated
behind `cfg(unix)` and the module did not compile there at all. Both the
module and its wiring are now gated, with the non-unix branch verified to
compile by forcing it locally rather than assumed.

`github_proxy.rs` and `main.rs` both crossed the 1000-line limit: the proxy
tests move to their own file and `router providers` to `providers_cli`.

CI measured 81.405611% against the 81.360862% recorded here — a rise, so the
ratchet asks for its figure to be committed for review.
